requirements.txt
Django==1.6
django-crispy-forms==1.4.0
simplejson==3.5.3
